Fixed compile errors on rendersystems and nature plugin

+++ RenderSystems/Direct3D7/include/OgreD3D7RenderSystem.h
@@ -293,7 +293,7 @@ namespace Ogre {
         /** See
           RenderSystem
          */
-        void _setTextureCoordSet( size_t unit, int index );
+        void _setTextureCoordSet( size_t unit, size_t index );
         /** See
           RenderSystem
          */
