+++ RenderSystems/Direct3D9/include/OgreD3D9RenderSystem.h
@@ -183,7 +183,7 @@ namespace Ogre
 		void _setProjectionMatrix( const Matrix4 &m );
 		void _setSurfaceParams( const ColourValue &ambient, const ColourValue &diffuse, const ColourValue &specular, const ColourValue &emissive, Real shininess );
 		void _setTexture( size_t unit, bool enabled, const String &texname );
-        void _setTextureCoordSet( size_t unit, int index );
+        void _setTextureCoordSet( size_t unit, size_t index );
         void _setTextureCoordCalculation(size_t unit, TexCoordCalcMethod m);
 		void _setTextureBlendMode( size_t unit, const LayerBlendModeEx& bm );
 		void _setTextureAddressingMode( size_t unit, TextureUnitState::TextureAddressingMode tam );
