+++ RenderSystems/Direct3D9/src/OgreD3D9RenderSystem.cpp
@@ -1099,7 +1099,7 @@ namespace Ogre
 		}
 	}
 	//---------------------------------------------------------------------
-	void D3D9RenderSystem::_setTextureCoordSet( size_t stage, int index )
+	void D3D9RenderSystem::_setTextureCoordSet( size_t stage, size_t index )
 	{
 		HRESULT hr;
 		hr = __SetTextureStageState( stage, D3DTSS_TEXCOORDINDEX, index );
